Begin fix compiler errors, tests using TypeScript
Beginning to fix the many TypeScript errors so that we can actually take
advantage of the real errors the TypeScript compiler tells us.

When switching over a .js file over to .ts, there will be a lot of unexpected
properties, so it will take time to fix these errors.
