+++ src/bell/Badge.js
@@ -1,6 +1,3 @@
-import { isPushNotificationsSupported, getConsoleStyle, addCssClass, removeCssClass, once } from '../utils';
-import * as log from 'loglevel';
-import Event from '../events.js';
 import ActiveAnimatedElement from './ActiveAnimatedElement.js';
