+++ src/entry.ts
@@ -7,9 +7,9 @@ if (Environment.isBrowser()) {
   incrementSdkLoadCount();
   if (getSdkLoadCount() > 1) {
     log.warn(`OneSignal: The web push SDK is included more than once. For optimal performance, please include our ` +
-             `SDK only once on your page.`);
+      `SDK only once on your page.`);
     log.debug(`OneSignal: Exiting from SDK initialization to prevent double-initialization errors. ` +
-              `Occurred ${getSdkLoadCount()} times.`);
+      `Occurred ${getSdkLoadCount()} times.`);
   } else {
     // We're running in the host page, iFrame of the host page, or popup window
     // Load OneSignal's web SDK
