+++ src/errors/OneSignalError.js
@@ -1,0 +1,8 @@
+import * as ExtendableError from 'es6-error';
+
+
+export default class OneSignalError extends ExtendableError {
+  constructor(message) {
+    super(message);
+  }
+}
