--- src/errors/OneSignalError.ts
@@ -1,3 +1,0 @@
-import * as ExtendableError from 'es6-error';
-
-export default class OneSignalError extends ExtendableError { }
