--- test/integration-server/config.json
@@ -1,33 +1,0 @@
-/*
-  Purpose: Configuration file for OneSignal web SDK tests runner
-
-  Description:
-
-  The tests execute in a browser environment, but a NodeJS test server runs side by side to facilitate database access
-  and perform other commands (e.g. possibly mouse clicks in the future) that would not be possible inside of a
-  browser's sandbox environment.
-
-  This configuration file is first requested before the tests begin running.
-*/
-{
-  "server": {
-    "rails_host_http": "localhost:3000",
-    "rails_host_https": "localhost:3001",
-    // The host to OneSignal's database that contains the HTTP and HTTPS apps for testing
-    "database_host": "localhost:5432",
-    "database_user": "postgres",
-    "database_pass": "",
-    "master_database": "gamethrive_00",
-    "shard_1_database": "gamethrive_80",
-  },
-  "tests": {
-    "rails_host_http": "ubuntu:3000",
-    "rails_host_https": "ubuntu:3001",
-    // OneSignal user auth key to create, delete, and list apps
-    "user_auth_key": "OTU2MWYxYzItNzdmZi00MGM3LWE3ODQtMmU5OWM1NWIyMTdl",
-  },
-  "chrome_extension": {
-    // The Chrome extension ID, which stays consistent due to our setting the key property of the extension's manifest.json
-    "extension_id": "engiogdknecmkikekehomfbhnneekmng"
-  }
-}
