+++ test/integration-server/entry.js
@@ -1,0 +1,1 @@
+import './server';
