+++ test/integration/extension.js
