+++ test/integration/multiStepSoloTest.js
