--- test/sandbox/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Web Push Sandbox - OneSignal</title>
-
-</head>
-<body>
-
-</body>
-</html>
