+++ webpack.config.js
@@ -146,7 +146,7 @@ const ONESIGNAL_WEB_SDK = {
     },
       {
         test: /\.scss$/,
-        loaders: IS_PROD ? ["style", "css", "autoprefixer-loader", "sass"] : ["style", "css", "autoprefixer-loader", "sass"]
+        loaders: ["style", "css", "autoprefixer-loader", "sass"]
       }]
   },
   resolve: {
@@ -166,7 +166,7 @@ function getWebSdkTestModuleName() {
 
 const ONESIGNAL_WEB_SDK_TESTS = {
   target: 'web',
-  entry: ['babel-polyfill', './test/entry.js'],
+  entry: ['./test/integration/entry.ts'],
   output: {
     path: path.join(__dirname, 'dist'),
     filename: getWebSdkTestModuleName() + '.js'
@@ -174,26 +174,17 @@ const ONESIGNAL_WEB_SDK_TESTS = {
   devtool: 'source-map',
   module: {
     loaders: [{
-      test: /\.js$/,
+      test: /\.(t|j)s$/,
       include: [
         path.resolve(__dirname, "./src"),
-        path.resolve(__dirname, "./test")
+        path.resolve(__dirname, "./test/integration"),
       ],
-      exclude: /(node_modules|bower_components|test\/server)/,
-      loader: 'babel-loader',
-      query: {
-        presets: ['es2015', 'stage-0'],
-        plugins: ['transform-runtime'],
-        cacheDirectory: true
-      }
-    },
-      {
-        test: /\.scss$/,
-        loaders: IS_PROD ? ["style", "css", "autoprefixer-loader", "sass"] : ["style", "css", "autoprefixer-loader", "sass"]
-      }]
+      exclude: /(node_modules)/,
+      loader: 'awesome-typescript-loader'
+    }]
   },
-  sassLoader: {
-    includePaths: [path.resolve(__dirname, "./src")]
+  resolve: {
+    extensions: ["", ".ts", ".js"]
   },
   debug: !IS_PROD,
   plugins: [
@@ -213,7 +204,7 @@ function getWebSdkTestServerModuleName() {
 
 const ONESIGNAL_WEB_SDK_TEST_SERVER = {
   target: 'node',
-  entry: ['babel-polyfill', './test/server.js'],
+  entry: ['babel-polyfill', './test/integration-server/entry.js'],
   output: {
     path: path.join(__dirname, 'dist'),
     filename: getWebSdkTestServerModuleName() + '.js'
@@ -222,8 +213,8 @@ const ONESIGNAL_WEB_SDK_TEST_SERVER = {
   module: {
     loaders: [{
       test: /\.js$/,
-      include: [path.resolve(__dirname, "./test/server")],
-      exclude: /(node_modules|bower_components)/,
+      include: [path.resolve(__dirname, "./test/integration-server/entry.js")],
+      exclude: /(node_modules)/,
       loader: 'babel-loader',
       query: {
         presets: ['es2015', 'stage-0'],
@@ -235,6 +226,9 @@ const ONESIGNAL_WEB_SDK_TEST_SERVER = {
         test: /\.json$/,
         loader: "json-loader"
       }]
+  },
+  resolve: {
+    extensions: ["", ".ts", ".js"]
   },
   debug: !IS_PROD,
   plugins: [
